renaming payment method classes: omit suffix Data
