+++ src/Mapper/RequestMapper.php
@@ -34,14 +34,14 @@ namespace Wirecard\PaymentSdk\Mapper;
 
 use Wirecard\PaymentSdk\Config;
 use Wirecard\PaymentSdk\Exception\MandatoryFieldMissingException;
-use Wirecard\PaymentSdk\Entity\PaymentMethod\CreditCardData;
+use Wirecard\PaymentSdk\Entity\PaymentMethod\CreditCard;
 use Wirecard\PaymentSdk\Transaction\CancelTransaction;
 use Wirecard\PaymentSdk\Transaction\InitialTransaction;
-use Wirecard\PaymentSdk\Entity\PaymentMethod\PayPalData;
+use Wirecard\PaymentSdk\Entity\PaymentMethod\PayPal;
 use Wirecard\PaymentSdk\Transaction\PayTransaction;
 use Wirecard\PaymentSdk\Transaction\ReserveTransaction;
 use Wirecard\PaymentSdk\Transaction\ThreeDAuthorizationTransaction;
-use Wirecard\PaymentSdk\Entity\PaymentMethod\ThreeDCreditCardData;
+use Wirecard\PaymentSdk\Entity\PaymentMethod\ThreeDCreditCard;
 use Wirecard\PaymentSdk\Transaction\Transaction;
 
 class RequestMapper
@@ -181,7 +181,7 @@ class RequestMapper
 
         $specificProperties['ip-address'] = $_SERVER['REMOTE_ADDR'];
 
-        if ($transaction->getPaymentTypeSpecificData() instanceof ThreeDCreditCardData) {
+        if ($transaction->getPaymentTypeSpecificData() instanceof ThreeDCreditCard) {
             $threeDProperties = [
                 self::PARAM_TRANSACTION_TYPE => 'check-enrollment',
             ];
