+++ src/Mapper/ResponseMapper.php
@@ -43,8 +43,8 @@ use Wirecard\PaymentSdk\Response\FormInteractionResponse;
 use Wirecard\PaymentSdk\Response\InteractionResponse;
 use Wirecard\PaymentSdk\Response\Response;
 use Wirecard\PaymentSdk\Response\SuccessResponse;
-use Wirecard\PaymentSdk\Entity\PaymentMethod\CreditCardData;
-use Wirecard\PaymentSdk\Entity\PaymentMethod\ThreeDCreditCardData;
+use Wirecard\PaymentSdk\Entity\PaymentMethod\CreditCard;
+use Wirecard\PaymentSdk\Entity\PaymentMethod\ThreeDCreditCard;
 use Wirecard\PaymentSdk\Transaction\Transaction;
 
 /**
@@ -57,11 +57,11 @@ class ResponseMapper
      * map the xml Response from engine to ResponseObjects
      *
      * @param $xmlResponse
-     * @param CreditCardData $transaction
+     * @param CreditCard $transaction
      * @return Response
      * @throws MalformedResponseException
      */
-    public function map($xmlResponse, CreditCardData $transaction = null)
+    public function map($xmlResponse, CreditCard $transaction = null)
     {
         $decodedResponse = base64_decode($xmlResponse);
         $xmlResponse = (base64_encode($decodedResponse) === $xmlResponse) ? $decodedResponse : $xmlResponse;
@@ -214,7 +214,7 @@ class ResponseMapper
         return (string)$result;
     }
 
-    private function mapThreeDResponse($payload, $response, $status, ThreeDCreditCardData $transaction)
+    private function mapThreeDResponse($payload, $response, $status, ThreeDCreditCard $transaction)
     {
         if (!isset($response->{'three-d'})) {
             throw new MalformedResponseException('Missing three-d element in enrollment-check response.');
@@ -247,13 +247,13 @@ class ResponseMapper
      * @param $xmlResponse
      * @param $response
      * @param $statusCollection
-     * @param CreditCardData $transaction
+     * @param CreditCard $transaction
      * @return FormInteractionResponse|InteractionResponse|SuccessResponse
      * @throws MalformedResponseException
      */
-    private function mapSuccessResponse($xmlResponse, $response, $statusCollection, CreditCardData $transaction = null)
+    private function mapSuccessResponse($xmlResponse, $response, $statusCollection, CreditCard $transaction = null)
     {
-        if ($transaction instanceof ThreeDCreditCardData) {
+        if ($transaction instanceof ThreeDCreditCard) {
             return $this->mapThreeDResponse($xmlResponse, $response, $statusCollection, $transaction);
         }
